can now set reels profile
