+++ router/StackNavigation.jsx
@@ -71,7 +71,6 @@ const StackNavigation = () => {
                     <Screen name='UserProfile' component={UserProfile} />
                     <Screen name='ViewReel' component={ViewReel} />
                     <Screen name='Message' component={Message} />
-                    <Screen name='ViewAvatar' component={ViewAvatar} />
                     <Screen name='ViewVideo' component={ViewVideo} />
                     <Screen name='MessageCamera' component={MessageCamera} />
                     <Screen name='PreviewMessageImage' component={PreviewMessageImage} />
@@ -154,6 +153,16 @@ const StackNavigation = () => {
                     <Screen
                       name='ReelsOption'
                       component={ReelsOption}
+                      options={{
+                        ...TransitionPresets.FadeFromBottomAndroid,
+                        cardStyle: {
+                          backgroundColor: color.transparent
+                        }
+                      }}
+                    />
+                    <Screen
+                      name='ViewAvatar'
+                      component={ViewAvatar}
                       options={{
                         ...TransitionPresets.FadeFromBottomAndroid,
                         cardStyle: {
