+++ screens/modals/ViewAvatar.jsx
@@ -4,7 +4,6 @@ import { Dimensions,  ImageBackground } from 'react-native'
 import AutoHeightImage from 'react-native-auto-height-image'
 
 import { useRoute } from '@react-navigation/native'
-import color from '../../style/color'
 import { img } from '../../style/viewAssets'
 
 const ViewAvatar = () => {
